+++ marathon/generator_test.go
@@ -12,14 +12,14 @@ import (
 
 func TestServicesFromMarathon(t *testing.T) {
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == "GET" && r.RequestURI == "/v2/apps" {
+		if r.Method == "GET" && r.RequestURI == "/v2/apps" && r.Header.Get("Accept") == "application/json" {
 			marathonApps := Apps{
 				Apps: []App{
 					App{
 						Id: "/redis",
 						Container: Container{
 							Docker: Docker{
-								PortMappings: []PortMapping{PortMapping{Protocol: "tcp", ServicePort: 6379}},
+								PortMappings: []PortMapping{PortMapping{ContainerPort: 6379, Protocol: "tcp", ServicePort: 41000}},
 							},
 						},
 					},
@@ -27,7 +27,7 @@ func TestServicesFromMarathon(t *testing.T) {
 						Id: "/registry",
 						Container: Container{
 							Docker: Docker{
-								PortMappings: []PortMapping{PortMapping{Protocol: "tcp", ServicePort: 80}},
+								PortMappings: []PortMapping{PortMapping{ContainerPort: 5000, Protocol: "tcp", ServicePort: 42000}},
 							},
 						},
 					},
@@ -36,9 +36,9 @@ func TestServicesFromMarathon(t *testing.T) {
 						Container: Container{
 							Docker: Docker{
 								PortMappings: []PortMapping{
-									PortMapping{Protocol: "tcp", ServicePort: 80},
-									PortMapping{Protocol: "tcp", ServicePort: 2003},
-									PortMapping{Protocol: "udp", ServicePort: 8125},
+									PortMapping{ContainerPort: 80, Protocol: "tcp", ServicePort: 43000},
+									PortMapping{ContainerPort: 2003, Protocol: "tcp", ServicePort: 43001},
+									PortMapping{ContainerPort: 8125, Protocol: "udp", ServicePort: 43002},
 								},
 							},
 						},
@@ -55,13 +55,13 @@ func TestServicesFromMarathon(t *testing.T) {
 			return
 		}
 
-		if r.Method == "GET" && r.RequestURI == "/v2/tasks" {
+		if r.Method == "GET" && r.RequestURI == "/v2/tasks" && r.Header.Get("Accept") == "application/json" {
 			marathonTasks := Tasks{
 				Tasks: []Task{
-					Task{AppId: "/redis", Host: "10.10.10.10", Ports: []int{31001}, ServicePorts: []int{6379}},
-					Task{AppId: "/redis", Host: "10.10.10.10", Ports: []int{31003}, ServicePorts: []int{6379}},
-					Task{AppId: "/registry", Host: "10.10.10.10", Ports: []int{31002}, ServicePorts: []int{80}},
-					Task{AppId: "/graphite-statsd", Host: "10.10.10.11", Ports: []int{31001, 31002, 31003}, ServicePorts: []int{80, 2003, 8125}},
+					Task{AppId: "/redis", Host: "10.10.10.10", Ports: []int{31001}, ServicePorts: []int{41000}},
+					Task{AppId: "/redis", Host: "10.10.10.10", Ports: []int{31003}, ServicePorts: []int{41000}},
+					Task{AppId: "/registry", Host: "10.10.10.10", Ports: []int{31002}, ServicePorts: []int{42000}},
+					Task{AppId: "/graphite-statsd", Host: "10.10.10.11", Ports: []int{31001, 31002, 31003}, ServicePorts: []int{43000, 43001, 43002}},
 				},
 			}
 
@@ -80,9 +80,8 @@ func TestServicesFromMarathon(t *testing.T) {
 	c := &http.Client{}
 
 	generator := Generator{
-		httpClient:     c,
-		domainStrategy: func(id string) string { return id },
-		config:         &Config{Servers: ts.URL},
+		httpClient: c,
+		config:     &Config{Servers: ts.URL},
 	}
 
 	services, _ := generator.Generate()
@@ -92,39 +91,49 @@ func TestServicesFromMarathon(t *testing.T) {
 	require.Len(t, services, 5)
 
 	require.Equal(t, services[0].Id, "/redis")
-	require.Equal(t, services[0].Domain, "/redis")
-	require.Equal(t, services[0].ServicePort, 6379)
+	require.Equal(t, services[0].Domain, "")
+	require.Equal(t, services[0].Port, 6379)
 	require.Equal(t, services[0].Protocol, "tcp")
+	require.Equal(t, services[0].ServicePort, 41000)
+	require.Equal(t, services[0].Source, "Marathon")
 	require.Equal(t, services[0].Hosts[0].Ip, "10.10.10.10")
 	require.Equal(t, services[0].Hosts[0].Port, 31001)
 	require.Equal(t, services[0].Hosts[1].Ip, "10.10.10.10")
 	require.Equal(t, services[0].Hosts[1].Port, 31003)
 
 	require.Equal(t, services[1].Id, "/registry")
-	require.Equal(t, services[1].Domain, "/registry")
-	require.Equal(t, services[1].ServicePort, 80)
+	require.Equal(t, services[1].Domain, "")
+	require.Equal(t, services[1].Port, 5000)
 	require.Equal(t, services[1].Protocol, "tcp")
+	require.Equal(t, services[1].ServicePort, 42000)
+	require.Equal(t, services[1].Source, "Marathon")
 	require.Equal(t, services[1].Hosts[0].Ip, "10.10.10.10")
 	require.Equal(t, services[1].Hosts[0].Port, 31002)
 
 	require.Equal(t, services[2].Id, "/graphite-statsd")
-	require.Equal(t, services[2].Domain, "/graphite-statsd")
-	require.Equal(t, services[2].ServicePort, 80)
+	require.Equal(t, services[2].Domain, "")
+	require.Equal(t, services[2].Port, 80)
 	require.Equal(t, services[2].Protocol, "tcp")
+	require.Equal(t, services[2].ServicePort, 43000)
+	require.Equal(t, services[2].Source, "Marathon")
 	require.Equal(t, services[2].Hosts[0].Ip, "10.10.10.11")
 	require.Equal(t, services[2].Hosts[0].Port, 31001)
 
 	require.Equal(t, services[3].Id, "/graphite-statsd")
-	require.Equal(t, services[3].Domain, "/graphite-statsd")
-	require.Equal(t, services[3].ServicePort, 2003)
+	require.Equal(t, services[3].Domain, "")
+	require.Equal(t, services[3].Port, 2003)
 	require.Equal(t, services[3].Protocol, "tcp")
+	require.Equal(t, services[3].ServicePort, 43001)
+	require.Equal(t, services[3].Source, "Marathon")
 	require.Equal(t, services[3].Hosts[0].Ip, "10.10.10.11")
 	require.Equal(t, services[3].Hosts[0].Port, 31002)
 
 	require.Equal(t, services[4].Id, "/graphite-statsd")
-	require.Equal(t, services[4].Domain, "/graphite-statsd")
-	require.Equal(t, services[4].ServicePort, 8125)
+	require.Equal(t, services[4].Domain, "")
+	require.Equal(t, services[4].Port, 8125)
 	require.Equal(t, services[4].Protocol, "udp")
+	require.Equal(t, services[4].ServicePort, 43002)
+	require.Equal(t, services[4].Source, "Marathon")
 	require.Equal(t, services[4].Hosts[0].Ip, "10.10.10.11")
 	require.Equal(t, services[4].Hosts[0].Port, 31003)
 }
@@ -176,30 +185,11 @@ func TestShouldNotConsiderAppsWithoutPorts(t *testing.T) {
 	c := &http.Client{}
 
 	generator := Generator{
-		httpClient:     c,
-		domainStrategy: func(id string) string { return id },
-		config:         &Config{Servers: ts.URL},
+		httpClient: c,
+		config:     &Config{Servers: ts.URL},
 	}
 
 	services, _ := generator.Generate()
 
 	require.Empty(t, services)
-}
-
-func TestIdToDomainReverse(t *testing.T) {
-	domain := IdToDomainReverse("/redis")
-
-	require.Equal(t, "redis", domain)
-
-	domain = IdToDomainReverse("/com/example/redis")
-
-	require.Equal(t, "redis.example.com", domain)
-}
-
-func TestLastPartOfIdAndSuffix(t *testing.T) {
-	g := LastPartOfIdAndSuffix{suffix: "example.com"}
-
-	domain := g.ToDomain("/group/service")
-
-	require.Equal(t, "service.example.com", domain)
 }