+++ mesos_master/generator.go
@@ -24,11 +24,12 @@ func (m *MesosMasterServiceGenerator) Generate() ([]types.Service, error) {
 	}
 
 	service := types.Service{
-		Domain:      m.config.Domain,
-		Hosts:       []types.Host{host},
-		Id:          "/mesos-master",
-		Protocol:    "tcp",
-		ServicePort: 80,
+		Domain:   m.config.Domain,
+		Hosts:    []types.Host{host},
+		Id:       "/mesos-master",
+		Port:     80,
+		Protocol: "tcp",
+		Source:   "Mesos Master",
 	}
 
 	return []types.Service{service}, nil