+++ mesos_master/generator_test.go
@@ -41,8 +41,9 @@ func TestShouldGenerateOneService(t *testing.T) {
 
 	require.Equal(t, services[0].Id, "/mesos-master")
 	require.Equal(t, services[0].Domain, domain)
-	require.Equal(t, services[0].ServicePort, 80)
+	require.Equal(t, services[0].Port, 80)
 	require.Equal(t, services[0].Protocol, "tcp")
+	require.Equal(t, services[0].Source, "Mesos Master")
 	require.Equal(t, services[0].Hosts[0].Ip, "10.10.10.10")
 	require.Equal(t, services[0].Hosts[0].Port, 5050)
 }